requirements.txt
Django>=3.0,<4.0
mongoengine==0.23
Pillow
djangorestframework>=3.1,<4.0
django-rest-framework-mongoengine>=3.4,<5
blinker==1.4